[D2M] Promote DstStackAllocator to Shared.{h,cpp}
Move the class declaration and implementation out of `Scheduled.cpp`
and into the shared header/source so the unscheduled pass can adopt it
in the next commit.  No behavior change: the API surface and the
slot-allocation order are identical, and all
`insert_dst_register_access_*.mlir` lit tests pass unchanged.

`Scheduled.cpp` no longer needs `STLExtras.h` / `Debug.h` /
`DebugLog.h` / `<deque>` directly; trim them.

Preparatory for #8364.
