+++ lib/Dialect/D2M/Transforms/InsertDstRegisterAccess/Scheduled.cpp
@@ -16,11 +16,6 @@
 #include "mlir/Dialect/MemRef/IR/MemRef.h"
 #include "mlir/Dialect/SCF/IR/SCF.h"
 #include "mlir/Transforms/GreedyPatternRewriteDriver.h"
-#include "llvm/ADT/STLExtras.h"
-#include "llvm/Support/Debug.h"
-#include "llvm/Support/DebugLog.h"
-
-#include <optional>
 
 #define DEBUG_TYPE "D2MInsertDstRegisterAccessScheduled"
 
@@ -31,183 +26,6 @@ namespace mlir::tt::d2m {
 namespace {
 
 using namespace detail;
-
-// ---------------------------------------------------------------------------
-// Stack-based allocator used by the scheduled path.
-//
-// Slots come from a free list (`sliceStack`) and are bound to one of four
-// explicit roles:
-//
-//   - `inputStack`     : operand reads of the in-progress compute op, popped
-//                        LIFO (most recent input is deallocated first).
-//   - `currentOutput`  : the result tile of the most recently emitted
-//                        compute op.  Its value may still be consumed
-//                        in-place by the next compute op.
-//   - `retiredOutputs` : prior outputs whose consumer has already been
-//                        emitted; they are dead but not yet recycled.
-//   - `scratchSlots`   : per-op private scratch (e.g. SFPU int reductions
-//                        via `getNumDstScratchSlices()`).  Owned by the op
-//                        for the lifetime of the region; never recycled,
-//                        and deliberately not tracked by `inputStack`,
-//                        `currentOutput`, or `currSliceIndex`.
-//
-// `currSliceIndex` is the most recently allocated *operand or output* slot
-// (never a scratch slot) so callers wanting to overwrite-in-place can grab
-// it directly.
-// ---------------------------------------------------------------------------
-
-class DstStackAllocator {
-public:
-  DstStackAllocator() = delete;
-  explicit DstStackAllocator(unsigned dstSliceCapacityIn)
-      : dstSliceCapacity(dstSliceCapacityIn) {
-    initSliceStack();
-  }
-
-  unsigned allocate(bool isStore = false);
-  unsigned allocateScratch();
-  unsigned deallocate();
-  void setStoreToDst() { storedToDst = true; }
-  bool didStoreToDst() const { return storedToDst; }
-  unsigned getCurrSliceIndex() const { return currSliceIndex; }
-  unsigned getFirstInputSliceIndex() const;
-  void deallocateAllButFirstInput();
-
-private:
-  unsigned dstSliceCapacity = 0;
-  unsigned currSliceIndex = 0;
-  SmallVector<unsigned, 16> inputStack;
-  std::optional<unsigned> currentOutput;
-  SmallVector<unsigned, 4> retiredOutputs;
-  SmallVector<unsigned, 4> scratchSlots;
-  SmallVector<unsigned, 16> sliceStack;
-  bool storedToDst = false;
-
-  void initSliceStack();
-};
-
-static void
-debugDumpDstStackAllocator(StringRef header, ArrayRef<unsigned> sliceStack,
-                           ArrayRef<unsigned> inputStack,
-                           std::optional<unsigned> currentOutput,
-                           ArrayRef<unsigned> retiredOutputs,
-                           ArrayRef<unsigned> scratchSlots,
-                           std::optional<unsigned> action) {
-  LDBG_OS([&](raw_ostream &os) {
-    os << header << "\n";
-    os << "  SliceStack     = ";
-    llvm::interleaveComma(sliceStack, os);
-    os << "\n  InputStack     = ";
-    llvm::interleaveComma(inputStack, os);
-    os << "\n  CurrentOutput  = ";
-    if (currentOutput) {
-      os << *currentOutput;
-    } else {
-      os << "(none)";
-    }
-    os << "\n  RetiredOutputs = ";
-    llvm::interleaveComma(retiredOutputs, os);
-    os << "\n  ScratchSlots   = ";
-    llvm::interleaveComma(scratchSlots, os);
-    if (action) {
-      os << "\n  --> " << *action;
-    }
-  });
-}
-
-unsigned DstStackAllocator::allocate(bool isStore) {
-  TT_assertv(!sliceStack.empty(), "Out of dst slices");
-
-  unsigned id = sliceStack.pop_back_val();
-  currSliceIndex = id;
-
-  if (isStore) {
-    // The previous `currentOutput`, if any, has no live consumer at this
-    // point (the new store supersedes it); demote it to `retiredOutputs`
-    // so a later `deallocate()` can recycle it cleanly without having to
-    // peek at queue tails.
-    if (currentOutput.has_value()) {
-      retiredOutputs.push_back(*currentOutput);
-    }
-    currentOutput = id;
-  } else {
-    inputStack.push_back(id);
-  }
-
-  debugDumpDstStackAllocator("== ALLOCATE ==", sliceStack, inputStack,
-                             currentOutput, retiredOutputs, scratchSlots, id);
-  return id;
-}
-
-unsigned DstStackAllocator::allocateScratch() {
-  TT_assertv(!sliceStack.empty(), "Out of dst slices");
-
-  unsigned id = sliceStack.pop_back_val();
-  scratchSlots.push_back(id);
-
-  // Intentionally do NOT update `currSliceIndex`, `inputStack`, or
-  // `currentOutput`.  Scratch is owned by the op for the lifetime of the
-  // region; it must not show up as a candidate for in-place reuse by
-  // later compute ops (the failure mode tracked in #8081).
-  debugDumpDstStackAllocator("== ALLOCATE SCRATCH ==", sliceStack, inputStack,
-                             currentOutput, retiredOutputs, scratchSlots, id);
-  return id;
-}
-
-unsigned DstStackAllocator::deallocate() {
-  TT_assertv(!(inputStack.empty() && retiredOutputs.empty() &&
-               !currentOutput.has_value()),
-             "Deallocating non-existent dst slice");
-
-  // Recycle in liveness order: most recent input first (LIFO), then any
-  // retired prior outputs, then the current output as a last resort
-  // (only legal once the next consumer is gone).  No middle-erase.
-  unsigned id = 0;
-  if (!inputStack.empty()) {
-    id = inputStack.pop_back_val();
-  } else if (!retiredOutputs.empty()) {
-    id = retiredOutputs.pop_back_val();
-  } else {
-    id = *currentOutput;
-    currentOutput.reset();
-  }
-
-  sliceStack.push_back(id);
-
-  debugDumpDstStackAllocator("== DEALLOCATE ==", sliceStack, inputStack,
-                             currentOutput, retiredOutputs, scratchSlots, id);
-  return id;
-}
-
-unsigned DstStackAllocator::getFirstInputSliceIndex() const {
-  TT_assertv(!inputStack.empty(), "No input slots allocated");
-  return inputStack.front();
-}
-
-void DstStackAllocator::deallocateAllButFirstInput() {
-  TT_assertv(inputStack.size() >= 1u, "Need at least one input to keep");
-
-  unsigned firstInput = inputStack.front();
-  inputStack.erase(inputStack.begin());
-
-  while (!inputStack.empty()) {
-    unsigned id = inputStack.pop_back_val();
-    sliceStack.push_back(id);
-    debugDumpDstStackAllocator("== DEALLOCATE (keeping first) ==", sliceStack,
-                               inputStack, currentOutput, retiredOutputs,
-                               scratchSlots, id);
-  }
-
-  currSliceIndex = firstInput;
-}
-
-void DstStackAllocator::initSliceStack() {
-  TT_assert((dstSliceCapacity > 0u && dstSliceCapacity <= 16u));
-
-  for (int i = dstSliceCapacity - 1; i >= 0; --i) {
-    sliceStack.push_back(static_cast<unsigned>(i));
-  }
-}
 
 // ---------------------------------------------------------------------------
 // Scheduled-only: in-place data copy generation (no loop cloning)
